DTO에 lombok 적용, 엔티티 키 Type을 Integer에서 Long으로 변경

+++ src/main/java/com/polychat/polychatbe/friend/command/domain/repository/FriendRepository.java
@@ -4,7 +4,7 @@ import com.polychat.polychatbe.friend.command.domain.model.Friend;
 import com.polychat.polychatbe.friend.command.domain.model.FriendUserId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FriendRepository extends JpaRepository<Friend, Integer> {
+public interface FriendRepository extends JpaRepository<Friend, Long> {
 
     public void deleteByUser1AndUser2(FriendUserId user1, FriendUserId user2);
 }
